+++ src/bundles.cpp
@@ -485,7 +485,7 @@ void HitBundle::combine(const vector<HitBundle*>& in_bundles,
     {
         int next_bundle = -1;
         const MateHit* next_hit=NULL; 
-        for(int i = 0; i < in_bundles.size(); ++i)
+        for(size_t i = 0; i < in_bundles.size(); ++i)
         {
             const vector<MateHit>& curr_hits = in_bundles[i]->hits();
